Bug 473862 - F5 key shortcut doesn't refresh project folder contents
